Sort out atom.xml for RSS feed
This sorts out the warning message about a missing layout and moves it to the root (as looking to phase out the blog directory).

It also ensures the site is set up for autodiscovery of the feed (should that still work anywhere!) and that a link is provided to the feed on each page.