Add creative commons license auto to all posts
The content of the site is different from the code of the site. As such I believe an appropriate license should be applied to each post to make it clear to anyone who wishes to use anything what they can do with the content. The MIT license for the code already covers the actual code elements.

Having used the tools available at https://creativecommons.org/choose/ this change adds a Creative Commons license notification to the posts layout so that each post automatically gets license details added to the footer.

It also updates the README to cover details about the use of the two types of licenses, and adds a plain text copy of the creative commons license to the `_posts` folder.