Add support for rich previews of the site
In many apps today like Twitter and Facebook when you post a link a rich preview of the site/page is provided. According to https://medium.com/@richardoosterhof/how-to-optimize-your-site-for-rich-previews-527ed13a6d69 to support this you need to add additional meta tags to the header of each page.

These include title, description and a link to an image. This changes adds the additional tags to the default layout, along with a new 'description' property to each existing pages front matter. This will then allow the population of the title and description to be dynamic depending on the page loaded.