Add post about adding support for rich previews
This post covers how I learnt that rich preview support for sites (when you post to twitter it shows a nice box with a description, image and title for your page, rather than just the raw url) was something you needed to provide support for in your own site, and not something that magically happened.

Also covers how I went about implementing it for my own site.