Fix missing OL style/code highlighting
Found that our default styling did not cover ordered lists (OL) so
corrected that.

Also felt that it wasn't as clear as it could be to differentiate
between the code blocks and the content of the post. Therefore tweaked
the highlight color to light grey to make the distinction clearer.