Add post on rel. between containers and the o/s
In a recent EA dev meetup we had a demo of Docker. One of the things that came up was why do we include a O/S in the `Dockerfile` when the documentation makes claims about Docker containers not needing an O/S.

No one could answer at the time, so I took it on myself to go away and do some digging. The result is this blog post.