Fix htmlproofer check failing in travis (#41)
Having re-read the documentation for htmlproofer the command it suggests running for Jekyll based projects is `htmlproofer --assume-extension ./_site`

The key bit missing from my command `--assume-extension`. From the docs this means

> Automatically add extension (e.g. `.html`) to file paths, to allow extensionless URLs (as supported by Jekyll 3 and GitHub Pages)

I also spotted a difference in how I had set the permalink attribute in the docker for development post (the one that always failed before) which led Jekyll building it the same as the other posts.